Create Readme.md
we will learn what is private and public cloud. Why companies are moving to public cloud, what are the advantages of moving to cloud.

Also, we will be introduced to the basics of AWS, including the core services and their significance in DevOps practices. Finally learn how to set up an AWS account and navigate the AWS Management Console.